refactor(web): move useScrollProgress hook to reusable location
Made-with: Cursor
